refactor: 사용자 평점 UI를 컴포넌트로 분리
